Added parameter files for the PREX sub-bank data structures.
Beginning with run 4747, the PREX data stream from the injector and
counting house uses sub-banks in each of the ROC banks.
The beamline and helicity map files have been modified to decode that
data beginning with that run number.

I have also added a "prex.conf" configuration which combines the
decodong using the CHA and injector detector maps.  This should
be expanded as we define the spectrometer data structures.