Fixed two missing initializations and removed data exchange requests used by mock data.
The scaler channels should be initialized to not require a refreence clock, unless one is specified in the channel map.
The helicity decoder bit counter was not explicity set to zero on initialization, and if it happened to evaluate to greater than the psuedorandom generator array, the predictor can get stuck off.
Both of these are now fixed.

I also have temporarily commeted out a series of external data exchange requests in the MollerMainDetector class whcih are needed for mock-data generation.  Thos ought to be shifted into the mock-data function within the  MollerMainDetector class, and will be later.