More parameter file updates for the PREX data files since run 1000.
The SAM and maindetector subsystems needed to have an eventcuts file
to set them in cut mode "3", and the CHA beamline eventcuts file
needed to be changed to put it in that mode too.
Corrected the VQWK number of samples expected from the injector (16564)
and the LHRS crate (8000) in the runs beginning with run 1000.