Modifications to the merge_from_QwAna_2 to produce matching results compared to master and develop branches
With these changes, the mock data generator has been verified to produce a datafile which matches that produced on "master" (commit  5a5d0d4), and the analysis of that data file produces runningsum outputs which match the analysis on the "master" branch.

The analysis of PREX test run 1098 produces runningsum outputs which match the analysis on the "develop" branch (commit 5a5d0d4).

The changes to QwHelicity are very similar to changes which had been made in develop.