Changed CHA beam current monitor names and added rough calibration
We realized that the channels previously labeled "bcm_us" and "bcm_ds"
were the unity-gain analog signals, and we didn't have the DSx3 and
DSx10 signals decoded.
We've changed these names to be "bcm_an_*" to be the analog BCMs.
The BCM DSx3 has been aliased as "q_targ".

Rough calibrations for pedestal and gain were entered for US, DS,
DSx3, and DSx10 based on 2uA nd 5uA data in run 1224.

An event cut on bcm_an_ds3 has been applied of >1uA for good events,
with a stability threshhold of 0.25.