Used "git update-index --chmod=-x <various files>" to clear the executable bit as described in issue #4
I repeatedly used a command like "find <some path>  -type f -executable" to find the files with executable bits set, and executed the git update-index command on them.
There remain several files in "Feedback" and "Tests" which have the executable bit, but I think they ought to.
There are several executables in "SetupFiles" which I think are no longer needed and could be removed from the distribution.  That's for another day.