Changed assert() statements
The assert() calls in the two scripts were not working correctly. The argument inside them was not being run so none of the regression values were being filled. The regression functions now run before the assert functions, and their return values are sent to the assert function using the testval variable.

 Please enter the commit message for your changes. Lines starting