Encloded bit shift operators in parenthases to clairify order of operation.
This addresses issue #63, in which confusing operator precedence was pointed out.
The primary fixes were in the routines realted to marker-word decoding in
QwEventBuffer and VQwSubsystem.  Likely the confusing precedence didn't matter
in the test case, because in the marker-word usage so far, the bank ID was
forced to zero.  But it should be clearer now.

Also changed the initalization order of one object in the QwEventBuffer
constructor to eliminate a compilation warning message.