Modified the tree-trim definition for "online" ROOT file production.
In the "online" mode root file production, the data processing performance
degrades as the root file increases in size.  By reducing the number of
data words stored in the trees, we can maintain functional performance
levels for a larger number of events.

The tree-trim file here will allow the "hw_sum" and "Device_Error_Code"
leaves from all of the standard beamline devices, and the SAMs to be
put into the evt and mul trees.  The event processing rate seems to
be sufficient to take 30 Hz data, but further trimming would be needed
for 240 Hz running.