Work In Progress:  adding a calcualtion of pair yields and asymmetries.
The status of completing a pair and having a pair asymmetry calculated
is done just before the main pttern completion is checked.  Right now
we're using an independent set of variables to hold the pair yield,
difference, and asymmetry.

This has been verified to compile, but I have not done any testing on it,
so it may not be doing what we meant for it to be doing.

Ready to hand this off to @leafybillow for testing and further development.