Cleaning up the pair tree and adding the "actual_helicity" flag to pattern and pair trees.
The "actual_helicity" flag in the pair tree is +1 for the first event being Hel+, and 0 for the first event being Hel-.  In the multiplet tree, that flag is always +1.

The pair tree will have the "yield" and "asym" values from the pair (note that "asym" for X and Y values are actually differences).

Minor correction of an error message that appears during pair analysis while building the helicity seed.