Support for ConstructObjects
In similar vein (and maybe in vain too) to the ConstructHistograms and
ConstructBranchAndTreeVector this provides another interface to the ROOT
file, but it is only supposed to be called once. Probably that means at
the beginning or at the end of the ROOT file, depending on what you use
it for (beginning if you write configuration and want to ensure data is
saved also when run crashes; end if you write summary information for
the entire run, like running sums etc).