Adding a summary of the number of good entries found by the correlator to its output.
The correlator will only accumulate entries if all of the data elements have a zero error code.
This change will count the number of entries with an error in each of the IVs and DVs, the total
number of entries passed to the correlator, and the number of good entries.
These counters are printed at the conclusion of the event loop, when the correlator would
attempt to perform the matrix inversions to determine the slopes.