Return an actual int in functions returning int [-Wreturn-type]
This actually causes crashes with gcc-8.3 when a calling function
directly uses the returned value itself in a return. In particular:
```

class A {
  public:
    int f(int i) { return this->f(i,i); };
    virtual int f(int,int) { return 0; };
};

class B: public A {
  public:
    int f(int,int) { /* ERROR HERE */ };
};

int main() {
  B* b = new B();
  A* a = b;
  std::cout << a->f(1) << std::endl;
}
```
when compiled as `g++ -o test -O2 test.cpp` (with warnings!) crashes with
```
Segmentation fault (core dumped)
```
Something a tiny bit more involved happens in japaan. I suspect this is
because of the different behavior of temporaries when returned (constexpr)
since the caller thinks it can evaluate the return value at compile time,
but in practice it fails since there is no temporary to return.

tl;dr
Don't ignore the warnings, please! This is why we can't have nice things.