WIP:  initial work to add the burp cut.
Ive added a calling routine for the burp cut to find a "burp"
and to mark the failng events.

THe next step would be to add a function to the data element
class (especailly the VQWK channel) to report a burp failure,
and methods to tell a channel what the burp cut parameters would
be.

The extent of the burp evaluation should be defined by a option
of the event ring, and the burp pre-cut should also be defined
as an option of the event ring.  The burp holdoff might be defined
in the event ring or in the data channels.