WIP  Still working on the burp cut.
Major changes are to have implemented a CheckForBurpFail in the
VQwHardwareChannel, and to have started the cascade for that
function through the composite classes and subsystems.

I think we don't need a special version in the classes derived
from VQwHardwareChannel (VQWK, scalers, etc).

I think we're not yet actualy calling the burp test function though.