Deaccumulation corner cases for VQWK channels
When deaccumulating with n2 < 0 (negative number of good events), in
particular n2 = -1, we have two special cases:

n2 = -1 leads to n = n1 + n2 = 1 (e.g. event ring has single event)
Here avg is non-zero but M2 is supposed to be zero, but numerical
stability can cause it to be slightly negative, and that fails the error
calculation (sqrt). So, compare M2 with
std::numerical_limits<double>::epsilon and round to zero if appropriate.

n2 = -1 leads to n = n1 + n2 = 0 (e.g. event ring is empty)
Here both avg and M2 are supposed to be zero, so compare with
std::numerical_limits<double>::epsilon and round to zero if appropriate.

With these tweaks NANs can only creep in when there are real errors, not
when there's merely rounding issues.

Good news: event ring properly unwinds VQWKs for 1296.