In online mode, allow interruption of the event loop by an "End" event or USR1 signal.
Added another signal handler to the EventBuffer to catch the USR1
signal, and exit the event loop, while in online mode.  If not
running in online mode, the USR1 signal has no effect.

There is now an option to have the event loop exit if the End" event
is caught by the analyzer.

In online mode, either of these exit events will result in the event
loop restarting as if we had killed and restarted the entire engine.