Another modification to the BMOD event cuts system.
I found that for my usual test run 1297, all events failed the
static BMOD cuts because the bmwobj was fixed at a non-zero value.
So, I have added a way to specify the upper and lower limits for
the bmwobj value for acceptable events.  If the upper limit is
less than the lower limit, the cut will be disabled.
If the bmwobj cut is enabled, it is always a global cut.

I added an empty bmod event cut file for run numbers below 1463.