Added a "event cut validity" test to the ADC18 class.
The event cut limits are initialized to -1 for the upper limit
and +1 for the lower limit.
One of the event cut methods tested for UL>LL before doing the
cut, but the other just tested them to be non-zero.  I've now
used the UL>LL test in both locations before attempting to
use them to mark events as failing.