Write the regression correlations and slopes to a tree
Instead of somehow working this into the existing framework,
it is a lot easier to store this in the blue....root files.

Since most of the stuff in the tree is a matrix, you may not
want to just double-click on the elements. Nothing happens.

Instead, you have to use something like:
lrb->Draw("A.fElements[0]")