changed accumulate and deaccumulate to accept additional arguments for error flags, and event count.
changed all instances of AccumulateRunningSum, and DeaccumulateRunningSum to accept an event count and errorflag. currently the errorflag argument is just used as a boolean-ish check to ignore other error flags for the burp cut calculation. I tested the outputs of this commit compared to a fresh pull of feature-burp-cut. the outputs match, so it seems the change has not affected how other sums are being accumulated.