Rough calibration for the cavity BPM Q cavities using runs 2793-2797.
The calibration for the cavity Q signals was done by finding the offset
in Q which seemed to give a "good" intersection of the XI vs Q trajectories
for data with a nominal X position of +2mm and -2mm.  It is not the best
fit of the Q values versus the Faraday cup currents measured during these
runs.
Maybe in the future, we should have one calibration of the Q cavity
to give the beam current, and also two independent calibrations to
be used for the X=XI/Q' and Y=YI/Q'' results.