Updating the main detector map to not be normalized during commissiong run with 5 uA beam for determining appropriate high voltages and placement
We also look at the SAMs, but those maps are unperturbed.

I'm also committing the pedestals we calculated for this specific voltage
at 5uA run and that file should be closed when we are done, but I'm committing
it for now to be safe.

I'm committing the online_shiftcrew.cfg that I have been editting a lot recently
and KK's suggestions of things to plot to look for correlations and problems in
the signal