Only close ROOT file if it has been opened successfully
It was possible to create a QwCorrelator data handler, but not fully
initialize it (e.g. the datahandlerarray_evt which isn't used at all yet
in QwParity doesn't ever call its ConstructTreeBranches method). This
resulted in null TFile pointer which is attempted to close. Result =
segfault on close.

Snuck in two other bug fixes which I thought might be at the cause:
- first close TFile, then delete histograms to make sure they make it
into the file
- only initialize histograms when disable-histos = false in the map file