When checking histogram/tree prefix, allow stuff after diff_
This ensures that "diff_|stat" is still treated as a diff_ for the
purpose of histogram/tree filling.

This is a bug fix for enabling pattern and burst histograms and trees,
since the type of histogram filling that is to be done changes between
constuction and filling (due to the use of "|stat"). In general, it may
have been a better idea to not reuse the same flag for histograms and
trees.