LRB Corrector: read in all cycles (i.e. bursts), pick cycle == 1
This maintains the functionality for the mulc_lrb tree, but adds the
mulc_lrb_burst tree with corrected value by the slopes of the first
burst.

This is a proof of concept: we can read in all slopes, but we have no
way yet to apply the appropriate one upon correction.