Reorganization in the publishable mixin class.
Added a "child" version of the mixing class, so that the same
function names can be specified in the template for both the
child and parent, yet have different functionality.  Maybe could
have (and maybe will) be done differently.

General cleanup of the publishable calls throughout exisiting
subsystems, now the "Request*" calls are thr preferred usage.

Removed the "friend" version of the ReturnInternalValue call,
and removed the declaration of friendship for the data handlers.

This complies on the adaq cluster, but has not been tested there,
or anywhere else.