Resolved two segfaults related to multiple datahandlerarray instances.
First, we have three instances of the datahandler arrays (event,
patterns-in-the-full-run, and patterns-in-a-burst), and they all
use the same map file to define their handlers.  However, most of
our handler instances to date are only meaningfully defined at one
of the pattern levels, and lead to incomplete initialization (and
pointers to nowhere) when instantiated in the event-based array.
I have added a scope marker to the array, which is filled
at creation based on the type of data object used to create the
array.  A new keyword "scope" would be added to the datahandler
mapfile entries to indicate if an instance should be created in the
"event" array, or the "pattern" array.

Second, the datahandler array histograms were created by using the
QwRootFile version of the ConstructHistograms function, which is
aware of the global "disable-histograms" flag, but were filled
by the QwDataHandlerArray version of FillHistograms, whcih is not
aware of it.  Thus, we could try to fill histograms which did not
exist (and segfault).
Modified the calls to FillHistograms to use the QwRootFile version.