A glorified amalgamation of the code in the invertMatrix.C bmod analysis code. This one still gives the same outputs, but it does it faster and with customizable input.txt config files.
New improvements:
There is a stub for putting your own cut on what counts as an acceptable slope (currently !=0).
The sensitivity data is stored in the ROOT file as well.
The slope file is maintained in an automated way, so re-analyzing a cycle will overwrite the prior instance of that cycle

Next stop: adding the plot making tools in as another script that reads the sensitivity/slope ROOT file as input