Added "alias-file-suff" values to the subblock correlator objects.
In the standard prex_datahandlers.map, I have added suffix values for the alias output files for the subblck correlators.  This results in five copies of the correaltion alias output files, one for the hw_sum, and then four for the subblocks.

This resolves issue #253 in the operations branch; although perhaps the run-ranged data handler files also need this fix.