Getting the running sum calculation for the data handlers to work.
The correlator now has a cycle counter that increments each time
we write a set of outputs.  This is independent of the burst
counter, but for the "burst correlator" does increment when the
burst counter does.

The VQwDataHandler error flag pointer gets assigned at the same
time as the source pointer.

VQwDataHandler contains its own running sum; this provides some
consistency in behavior between the correlator and other data handlers,
although it is inconsistent with how running sums work for the
subsystem arrays and pattern object.