Changing to use "bcm_target" as the normalizing BCM for all of PREX2.
For most runs beginning at run 2721, the normailzing BCM is defined to be
bcm_target = (2*bcm_an_us + bcm_an_ds)/3.
The exceptions are run ranges 3346-3347, 3363-3365, and 3594-3738 which
has bcm_target = bcm_dg_ds.

This set of changes just has the simple changes in files that keep the same
name, and the file moves for "3347-" -> "3348-" and "3583-" -> "3594-3738".
Two new files are added, for "3346-3347" and "3363-3365".