Removed two INJ beamline pedestal files for CREX period.
These two files were added from the operations directory on the adaq cluster,
and are not consistent with the "blessed" pedestal file lists from Caryn.

prexinj_beamline_pedestal.5517-.map was removed because it was
identical to prexinj_beamline_pedestal.5356-.map.
prexinj_beamline_pedestal.6832-.map was identical to prexinj_beamline_pedestal.6831-.map
up to FC2, and used the results from 5268 for FC2->ILD.