Fixed two data handler "connect" routines to find other data handler variables.
This addresses issue #296.  This was first noticed when trying to use the new
"diff_bpmE" energy variable in the dithering corrector object; variables
marked as independent variables would not be searched for in the data handler
array.  This fix copies the behavior which was already present for the dependent
variables, in which the DH array is searched first.