Changed channel publication failure messages to "Debug" level
The channel publication failure messages inside the detector array classes
are somewhat extraneous, as they are triggered when a detector channel is
not in that array, even if it is in a different array.  Objects which request
a channel to be published should be reporting a failure anyway.

This resolves issue #271.