Corrected a mapfile naming error for CREX period and improved mapfile error diagnostics.
I found the mapfile prexCH_beamline_pedestal.5436-5472.map had two digits transposed in
its name such that is was for the nonsensical run range 5436 - 4572.  That was corrected
by inspecting the adjacent run-ranged map files to find which runs were incorrectly labeled.

Also improved the error message from QwParameterFile so that it includes the two run numbers
when the second run number is smaller than the first to help finding these sorts of errors.