Decreased the "singularity threshold" for the LRB matrix inversion
Changed the lower limit of the LRB correlation matrix determinant from a value of ten to
the negative power of the number of IVs to be instead be ten to the negative power of
twice the number of IVs.
I have not found a clear justification for choosing a particular value for this threshold.
This now gives a threshold value of 1e-10 for our std regression set, and the determinants
overnight were in the neighborhood of ~4e-6.

Also modified the console output to give a header line for each LRB correlation calculation,
so we can see in the log files which regression set may have problems with matrix inversion.