New feature:  Adding an event-level BCM difference and a burp cut on it.
I have added an event-level BCM difference in prexCH_beamline.7502-.map, and a demonstration burp cut on it in prexCH_beamline_eventcuts.8301-.map.

During initial testing, I realized that the event-level combination was not set up to support negative weights on the devices; I have modified the combined BCM parsing routines to use teh sum of the absolute values of the weights on each device as the denominator.  This results in the BCM difference being reported as (bcm_an_us - bcm_an_ds)/2.